[refactor] App: split ui.utils into ui.composehelper and ui.extensions
