+++ composeApp/src/androidMain/kotlin/com/rwmobi/kunigami/ui/composehelper/GetScreenSizeInfo.android.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.composehelper
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
