+++ composeApp/src/androidMain/kotlin/com/rwmobi/kunigami/ui/composehelper/ShouldUseDarkTheme.android.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.composehelper
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
