+++ composeApp/src/androidMain/kotlin/com/rwmobi/kunigami/ui/extensions/GenerateRandomLong.android.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.extensions
 
 import java.util.UUID
