+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/composehelper/GenerateGYRHueColorPalette.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.composehelper
 
 import androidx.compose.ui.graphics.Color
