+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/agile/AgileScreen.kt
@@ -47,9 +47,9 @@ import com.rwmobi.kunigami.ui.components.koalaplot.VerticalBarChart
 import com.rwmobi.kunigami.ui.destinations.agile.components.CountDownWidget
 import com.rwmobi.kunigami.ui.model.chart.RequestedChartLayout
 import com.rwmobi.kunigami.ui.theme.getDimension
-import com.rwmobi.kunigami.ui.utils.generateGYRHueColorPalette
-import com.rwmobi.kunigami.ui.utils.getPercentageColorIndex
-import com.rwmobi.kunigami.ui.utils.partitionList
+import com.rwmobi.kunigami.ui.composehelper.generateGYRHueColorPalette
+import com.rwmobi.kunigami.ui.extensions.getPercentageColorIndex
+import com.rwmobi.kunigami.ui.extensions.partitionList
 import io.github.koalaplot.core.style.LineStyle
 import io.github.koalaplot.core.util.toString
 import io.github.koalaplot.core.xygraph.HorizontalLineAnnotation
