+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/agile/components/CountdownDashboard.kt
@@ -33,8 +33,8 @@ import androidx.compose.ui.unit.dp
 import com.rwmobi.kunigami.ui.model.rate.RateTrend
 import com.rwmobi.kunigami.ui.theme.AppTheme
 import com.rwmobi.kunigami.ui.theme.getDimension
-import com.rwmobi.kunigami.ui.utils.drawArcSegment
-import com.rwmobi.kunigami.ui.utils.generateGYRHueColorPalette
+import com.rwmobi.kunigami.ui.composehelper.drawArcSegment
+import com.rwmobi.kunigami.ui.composehelper.generateGYRHueColorPalette
 import kunigami.composeapp.generated.resources.Res
 import kunigami.composeapp.generated.resources.agile_expires_in
 import kunigami.composeapp.generated.resources.agile_unit_rate
