+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/navigation/AppNavigationHost.kt
@@ -25,8 +25,8 @@ import com.rwmobi.kunigami.ui.destinations.tariffs.TariffsScreen
 import com.rwmobi.kunigami.ui.destinations.tariffs.TariffsUIEvent
 import com.rwmobi.kunigami.ui.destinations.usage.UsageScreen
 import com.rwmobi.kunigami.ui.destinations.usage.UsageUIEvent
-import com.rwmobi.kunigami.ui.utils.collectAsStateMultiplatform
-import com.rwmobi.kunigami.ui.utils.getScreenSizeInfo
+import com.rwmobi.kunigami.ui.composehelper.collectAsStateMultiplatform
+import com.rwmobi.kunigami.ui.composehelper.getScreenSizeInfo
 import com.rwmobi.kunigami.ui.viewmodels.AccountViewModel
 import com.rwmobi.kunigami.ui.viewmodels.AgileViewModel
 import com.rwmobi.kunigami.ui.viewmodels.TariffsViewModel
