+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/theme/Theme.kt
@@ -4,7 +4,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import com.rwmobi.kunigami.ui.utils.shouldUseDarkTheme
+import com.rwmobi.kunigami.ui.composehelper.shouldUseDarkTheme
 
 private val mediumContrastLightColorScheme = lightColorScheme(
     primary = primaryLightMediumContrast,
