+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/viewmodels/AccountViewModel.kt
@@ -20,7 +20,7 @@ import com.rwmobi.kunigami.domain.usecase.UpdateMeterPreferenceUseCase
 import com.rwmobi.kunigami.ui.destinations.account.AccountScreenLayout
 import com.rwmobi.kunigami.ui.destinations.account.AccountUIState
 import com.rwmobi.kunigami.ui.model.ErrorMessage
-import com.rwmobi.kunigami.ui.utils.generateRandomLong
+import com.rwmobi.kunigami.ui.extensions.generateRandomLong
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
