+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/viewmodels/TariffsViewModel.kt
@@ -13,7 +13,7 @@ import co.touchlab.kermit.Logger
 import com.rwmobi.kunigami.domain.usecase.GetFilteredProductsUseCase
 import com.rwmobi.kunigami.ui.destinations.tariffs.TariffsUIState
 import com.rwmobi.kunigami.ui.model.ErrorMessage
-import com.rwmobi.kunigami.ui.utils.generateRandomLong
+import com.rwmobi.kunigami.ui.extensions.generateRandomLong
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
