+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/ui/extensions/DoubleExtensionsKtTest.kt
@@ -5,8 +5,9 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.extensions
 
+import com.rwmobi.kunigami.ui.extensions.getPercentageColorIndex
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
