+++ composeApp/src/desktopMain/kotlin/com/rwmobi/kunigami/ui/composehelper/CollectAsStateMultiplatform.desktop.kt
@@ -5,15 +5,15 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.composehelper
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
+import androidx.compose.runtime.collectAsState
 import kotlinx.coroutines.flow.StateFlow
 import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.EmptyCoroutineContext
 
 @Composable
-expect fun <T> StateFlow<T>.collectAsStateMultiplatform(
-    context: CoroutineContext = EmptyCoroutineContext,
-): State<T>
+actual fun <T> StateFlow<T>.collectAsStateMultiplatform(
+    context: CoroutineContext,
+): State<T> = collectAsState(context)
