+++ composeApp/src/desktopMain/kotlin/com/rwmobi/kunigami/ui/composehelper/GetScreenSizeInfo.desktop.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.composehelper
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
