+++ composeApp/src/iosMain/kotlin/com/rwmobi/kunigami/ui/extensions/GenerateRandomLong.ios.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.utils
+package com.rwmobi.kunigami.ui.extensions
 
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.UByteVar
